backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
Markdown==3.7
packaging==25.0
psycopg==3.2.5
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
